General refactorings
