+++ osu.Game/Screens/Select/FilterCriteria.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using JetBrains.Annotations;
 using osu.Game.Beatmaps;
 using osu.Game.Rulesets;
 using osu.Game.Screens.Select.Filter;
@@ -51,6 +52,10 @@ namespace osu.Game.Screens.Select
             }
         }
 
+        /// <summary>
+        /// The collection to filter beatmaps from.
+        /// </summary>
+        [CanBeNull]
         public CollectionFilter Collection;
 
         public struct OptionalRange<T> : IEquatable<OptionalRange<T>>
